Remove tiles from public API
