+++ test/update.test.ts
@@ -346,56 +346,6 @@ test('updateData: works with empty diff', () => {
     }).not.toThrow();
 });
 
-test('updateData: invalidates tiles at deeper zoom', () => {
-    const initialData = {
-        type: 'FeatureCollection' as const,
-        features: [{
-            type: 'Feature' as const,
-            id: 'feature1',
-            geometry: {
-                type: 'Polygon' as const,
-                coordinates: [[
-                    [0, 0], [5, 0], [5, 5], [0, 5], [0, 0]
-                ]]
-            },
-            properties: {name: 'Original'}
-        }]
-    };
-
-    const index = new GeoJSONVT(initialData, {
-        updateable: true,
-        indexMaxZoom: 5,
-        indexMaxPoints: 0
-    });
-
-    const tileId = toID(5, 16, 16);
-
-    const tileBefore = index.tiles[tileId];
-    expect(tileBefore).toBeTruthy();
-    expect(tileBefore.numFeatures).toBe(1);
-
-    const updatedFeature = {
-        type: 'Feature' as const,
-        id: 'feature1',
-        geometry: {
-            type: 'Polygon' as const,
-            coordinates: [[
-                [0, 0], [10, 0], [10, 10], [0, 10], [0, 0]
-            ]]
-        },
-        properties: {name: 'Updated'}
-    };
-
-    index.updateData({add: [updatedFeature]});
-
-    const tileAfter = index.tiles[tileId];
-    expect(tileAfter).toBeUndefined();
-
-    const tileRegenerated = index.getTile(5, 16, 16);
-    expect(tileRegenerated).toBeTruthy();
-    expect(tileRegenerated.features[0].tags.name).toBe('Updated');
-});
-
 test('updateData: invalidates tiles with partial intersection', () => {
     const initialData = {
         type: 'FeatureCollection' as const,
@@ -433,66 +383,6 @@ test('updateData: invalidates tiles with partial intersection', () => {
     const tile = index.getTile(2, 3, 2);
     expect(tile).toBeTruthy();
     expect(tile.features.length).toBe(2);
-});
-
-
-test('updateData: does not invalidate unaffected tiles', () => {
-    const initialData = {
-        type: 'FeatureCollection' as const,
-        features: [
-            {
-                type: 'Feature' as const,
-                id: 'northwest',
-                geometry: {
-                    type: 'Point' as const,
-                    coordinates: [-90, 45]  // NW quadrant only
-                },
-                properties: {}
-            },
-            {
-                type: 'Feature' as const,
-                id: 'southeast',
-                geometry: {
-                    type: 'Point' as const,
-                    coordinates: [90, -45]  // SE quadrant only
-                },
-                properties: {}
-            }
-        ]
-    };
-
-    const index = new GeoJSONVT(initialData, {
-        updateable: true,
-        indexMaxZoom: 2,
-        indexMaxPoints: 0
-    });
-
-    const nwTileId = toID(1, 0, 0);
-    const seTileId = toID(1, 1, 1);
-
-    const nwTileBefore = index.tiles[nwTileId];
-    const seTileBefore = index.tiles[seTileId];
-
-    expect(nwTileBefore).toBeTruthy();
-    expect(seTileBefore).toBeTruthy();
-
-    const updatedFeature = {
-        type: 'Feature' as const,
-        id: 'northwest',
-        geometry: {
-            type: 'Point' as const,
-            coordinates: [-85, 40]  // NW different coordinate
-        },
-        properties: {}
-    };
-
-    index.updateData({add: [updatedFeature]});
-
-    const nwTileAfter = index.tiles[nwTileId];
-    expect(nwTileAfter).toBeUndefined();
-
-    const seTileAfter = index.tiles[seTileId];
-    expect(seTileAfter).toBe(seTileBefore);
 });
 
 test('updateData: invalidates and regenerates tiles at multiple zoom levels', () => {
@@ -557,44 +447,6 @@ test('updateData: invalidates and regenerates tiles at multiple zoom levels', ()
 
     expect(newZ7Tile.features[0].id).toBe('feature1');
     expect(newZ7Tile.features[0].tags.name).toBe('Updated');
-});
-
-test('updateData: invalidates tiles when feature is within the buffer edge', () => {
-    const initialData = {
-        type: 'FeatureCollection' as const,
-        features: [{
-            type: 'Feature' as const,
-            id: 'feature1',
-            geometry: {
-                type: 'Point' as const,
-                coordinates: [-45, 45] // inside tile 1-0-0
-            },
-            properties: {}
-        }]
-    };
-
-    const index = new GeoJSONVT(initialData, {
-        updateable: true,
-        indexMaxZoom: 1,
-        indexMaxPoints: 0
-    });
-
-    const tileId = toID(1, 0, 0);
-    index.getTile(1, 0, 0);
-    expect(index.tiles[tileId]).toBeTruthy();
-
-    const featureWithinBuffer = {
-        type: 'Feature' as const,
-        id: 'buffer-feature',
-        geometry: {
-            type: 'Point' as const,
-            coordinates: [2, 0] // feature within tile buffer edge
-        },
-        properties: {}
-    };
-
-    index.updateData({add: [featureWithinBuffer]});
-    expect(index.tiles[tileId]).toBeUndefined();
 });
 
 test('updateData: handles drill-down after update', () => {
@@ -872,7 +724,3 @@ test('updateClusterOptions: can toggle clustering from on to off and then back o
     expect(index.getClusterExpansionZoom(clusterId)).toBeGreaterThan(0);
     expect(index.getTile(0, 0, 0).features.some(f => (f.tags as ClusterProperties)?.cluster)).toBe(true);
 });
-
-function toID(z: number, x: number, y: number): number {
-    return (((1 << z) * y + x) * 32) + z;
-}
